Adding the Advertiser object.
The advertiser object needs to include all attributes for both request
and response objects as both the DSP and SSP are going to need to serialize
and deserialize the object in the same manner.  As such, the blocklist
attribute is included in the object to support the AdvertiserBlocklistResponse
object.

Also, validating behavior in the Advertiser object with respect to the
blocklist attribute being non-null in all states.