Adding comments to interfaces.
Also, added some dummy data into the client code based upon our integration
test.  The dummy data is found on the SSP side as well.  We might want to
look at putting this in a properties file that is shared... somewhere.

More code cleanup coming... Renaming echanges and making it the
supply-side platform.